Add dialog for gaussian noise

+++ mainwindow.h
@@ -82,7 +82,7 @@ private:
     QString file_path, file_name;
     int Ix,Iy,stage_num, total_stages, ksw = 0, ksh = 0, windowsize = 3;
     bool auto_save_enable = false, from_input_method = false;
-    float mean, deviation;
+    double mean, deviation;
     QGraphicsScene *scene;
 
 };
